# flutter/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Travel Memory Vault API",
    "description": "Travel Memory Vault - Personal travel journal with AI-powered search",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/trips/{trip_id}/metadata": {
      "post": {
        "tags": [
          "Metadata"
        ],
        "summary": "Create Trip Metadata",
        "description": "Create metadata for a trip.\n\n**Authentication:** Required\n\n**Permissions:** Trip owner only\n\n**Request Body:**\n- traveler_type: Array of traveler types (solo, couple, family, group)\n- age_group: Target age group (gen-z, millennial, gen-x, boomer)\n- travel_style: Array of travel styles (adventure, luxury, budget, cultural, relaxed)\n- difficulty_level: Overall difficulty (easy, moderate, challenging, extreme)\n- budget_category: Budget level (budget, mid-range, luxury)\n- activity_focus: Array of activity types (hiking, food, photography, nightlife, beaches)\n- is_discoverable: Whether trip can be found in public search (default: false)\n- tags: User-defined tags\n\n**Returns:**\nCreated trip metadata\n\n**Errors:**\n- 404: Trip not found\n- 403: User doesn't own this trip\n- 409: Metadata already exists (use PATCH to update)\n- 422: Invalid enum values",
        "operationId": "create_trip_metadata_api_v1_trips__trip_id__metadata_post",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripMetadataCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Metadata"
        ],
        "summary": "Get Trip Metadata",
        "description": "Get metadata for a trip.\n\n**Authentication:** Required\n\n**Permissions:**\n- Trip owner: Always allowed\n- Others: Only if trip is public or unlisted\n\n**Returns:**\nTrip metadata\n\n**Errors:**\n- 404: Trip or metadata not found\n- 403: Trip is private and user is not owner",
        "operationId": "get_trip_metadata_api_v1_trips__trip_id__metadata_get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Metadata"
        ],
        "summary": "Update Trip Metadata",
        "description": "Update metadata for a trip (partial update).\n\n**Authentication:** Required\n\n**Permissions:** Trip owner only\n\n**Request Body:**\nAll fields are optional. Only provided fields will be updated.\n\n**Returns:**\nUpdated trip metadata\n\n**Errors:**\n- 404: Trip or metadata not found\n- 403: User doesn't own this trip\n- 422: Invalid enum values",
        "operationId": "update_trip_metadata_api_v1_trips__trip_id__metadata_patch",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripMetadataUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Metadata"
        ],
        "summary": "Delete Trip Metadata",
        "description": "Delete metadata for a trip.\n\n**Authentication:** Required\n\n**Permissions:** Trip owner only\n\n**Returns:**\n204 No Content on success\n\n**Errors:**\n- 404: Trip or metadata not found\n- 403: User doesn't own this trip\n\n**Note:**\nTrip itself is NOT deleted, only its metadata.",
        "operationId": "delete_trip_metadata_api_v1_trips__trip_id__metadata_delete",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Bearer token from Supabase Auth",
              "title": "Authorization"
            },
            "description": "Bearer token from Supabase Auth"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TripMetadataCreate": {
        "properties": {
          "traveler_type": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traveler Type"
          },
          "age_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Group"
          },
          "travel_style": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Style"
          },
          "difficulty_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Level"
          },
          "budget_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Category"
          },
          "activity_focus": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Focus"
          },
          "is_discoverable": {
            "type": "boolean",
            "title": "Is Discoverable",
            "default": false
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "TripMetadataCreate",
        "description": "Trip metadata creation schema.\n\nInherits all fields from TripMetadataBase.\nUsed for POST /trips/{trip_id}/metadata endpoint.\n\nBusiness Logic:\n    - trip_id is taken from URL path parameter\n    - All fields are optional except is_discoverable (defaults to False)"
      },
      "TripMetadataResponse": {
        "properties": {
          "traveler_type": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traveler Type"
          },
          "age_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Group"
          },
          "travel_style": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Style"
          },
          "difficulty_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Level"
          },
          "budget_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Category"
          },
          "activity_focus": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Focus"
          },
          "is_discoverable": {
            "type": "boolean",
            "title": "Is Discoverable",
            "default": false
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "trip_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trip Id"
          },
          "quality_score": {
            "type": "number",
            "title": "Quality Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "quality_score",
          "created_at",
          "updated_at"
        ],
        "title": "TripMetadataResponse",
        "description": "Trip metadata response schema.\n\nAttributes:\n    trip_id: Trip UUID\n    traveler_type: Array of traveler types\n    age_group: Target age group\n    travel_style: Array of travel styles\n    difficulty_level: Overall difficulty\n    budget_category: Budget level\n    activity_focus: Array of activity types\n    is_discoverable: Whether trip can be found in public search\n    quality_score: System-calculated quality score (0-1)\n    tags: User-defined tags\n    created_at: Creation timestamp\n    updated_at: Last update timestamp\n\nConfig:\n    from_attributes: Enable ORM mode for SQLAlchemy models"
      },
      "TripMetadataUpdate": {
        "properties": {
          "traveler_type": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traveler Type"
          },
          "age_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Group"
          },
          "travel_style": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Travel Style"
          },
          "difficulty_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty Level"
          },
          "budget_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Category"
          },
          "activity_focus": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Focus"
          },
          "is_discoverable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Discoverable"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "type": "object",
        "title": "TripMetadataUpdate",
        "description": "Trip metadata update schema.\n\nAll fields are optional for partial updates.\n\nAttributes:\n    traveler_type: New traveler types\n    age_group: New age group\n    travel_style: New travel styles\n    difficulty_level: New difficulty level\n    budget_category: New budget category\n    activity_focus: New activity focus\n    is_discoverable: New discoverable setting\n    tags: New tags\n\nBusiness Rules:\n    - Only owner can update\n    - Only provided fields are updated\n    - quality_score cannot be updated (system-managed)"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}